First idea of code (incomplete)
Be able to find the rotation and translation between two views of the same image. The rotation is due to change in the car heading and translation is due to imprecise GPS position. No probabilistic study (should be introduce as a next step)
Test on smiley images (to be tested on two real radar images)
main.py is the main file
data.py is a proposition class for data structure